Add files via upload
Jeongwon works in Sunheon cafeteria. Today’s lunch menu is bossam. Bossam is very popular menu, but its quantity is limited. To make sure that everyone gets bossam, Jeongwon has to set a limit on the maximum number of pieces per person. Given a number of people coming to the cafeteria, desired amount of bossam for each person, and total amount of bossam, implement an algorithm to compute the maximum limit ℓ so that every person gets at least ℓ amount of bossam.
For instance, consider the case that 4 people are coming to the cafeteria having 3, 1, 6, 5
pieces as desired amount of bossam, respectively, and we have total 13 pieces of bossams. Then
setting ℓ = 4, we can give 3, 1, 4, 4 pieces of bossam to them.
In the first line of the input, the number t of test cases is given.
 For each test case, the inputs are given as follows.
– In the first line of each input, the number n of people coming to the cafeteria is given.
(3 ≤ n ≤ 10, 000)
– In the second line of each input, n desired amounts m1, . . . , mn of the people are
given. (1 ≤ mi ≤ 100, 000)
– In the third line of each input, the total amount T of bossam is given. (n ≤ T ≤
1, 000, 000, 000).