Gathering Score
The rule of competition is
following. The player is given a n × m grid, filled with numbers. The player starts at left-top
corner cell. Each time, the player can move 1 cell to the right or below. When the player gets to
the right-bottom corner cell, the competition ends and the sum of the numbers in the path of
the player is the score of the player. Implement an algorithm to compute the maximum possible
score Eunhyeong can acheive when the grid is given.
For instance, if following 5 × 5 grid is given, Eunhyeong can get the score of 21 by following
red path.
In the first line of the input, the number t of test cases is given.
 For each test case, the input is given as following.
– In the first line of a test case, the size of the grid n and m are given. (1 ≤ n, m ≤
5, 000)
– In the next following n lines, m numbers in each row of the grid is given, from top
to bottom.