Inserting operators
Given numbers 1 2 . . . n (n < 9), we can make different numbers by inserting operators in
between the numbers. For instance consider n = 3. If we insert a ‘+’ between 2 and 3, then we
have 1 2 + 3 = 15. If we insert a ‘+’ between 1 and 2, and ‘-’ between 2 and 3, then we have
1 + 2 − 3 = 0. Given n and T, implement an algorithm to compute the ways to make the target
number T by inserting ‘+’s and ‘-’s in between the numbers 1 2 . . . n.
 In the first line of the input, the number t of test cases is given.
 For each test case, n and T are given in a line, separated by space. (1 ≤ n ≤ 9)