Minimum Palindrome Subdivision
Palindrome is a word the reads the same as backwards as forwords, such as racecar, noon,
and madam. Given a string, minimum palindrome subdivision is a subdivison of the string into
palindromes with minimum size. Implement an algorithm to compute the size of a minimum
palindrome subdivision of a given string.
For instance, {‘B’, ‘ABACABA’} is a minimum palindrome subdivisions of ‘BABACABA’.
 In the first line of the input, the number t of test cases is given.
 For each test case, the input is given as following
– In the first line of a test case, the length n of the string S is given. (1 ≤ n ≤ 10, 000)
– In the second line of a test case, the string S is given. (S consists of n english
characters)