Heaviest Liquid
In the chemistry laboratory in SM university, there are several bottles of different liquids. One
of the students was wondering how heavy the liquid can be when she fills an empty flask using
those liquids. Given the volumes and weights of the liquid bottles and the volume of the flask,
implement an algorithm to compute the maximum weight liquid possible when she fills up the
flask. (We assume that the volumes and weights simply add up when mix two different liquids
are mixed.)
 In the first line of the input, the number t of test cases is given.
 For each test case, the input is given as following.
– In the first line of a test case, the number of liquids n is given. (1 ≤ n ≤ 1, 000, 000)
– In the second line, n volumes v1, . . . , vn of the liquids are given in a line. (1 ≤ vi ≤
20, 000)
– In the third line, n weights w1, . . . , wn of the liquids are given in the line. (1 ≤ wi ≤
10, 000)
– In the last line, the volume p of flask is given. (1 ≤ p ≤ 100, 000, 000)