Classroom Assignment
Hermoine has now become a graduate school student. One of her job is to assign a classroom
to each of the lectures held during this semester. Of course, she cannot assign two lectures with
overlapping lecture time to the same classroom. Given the start times and end times of the
lectures, implement an algorithm to compute the minimum number of classrooms needed to
assign a classroom to every lecture.
 In the first line of the input, the number t of test cases is given.
 For each test case, the input is given as following.
– In the first line of a test case, the number of lectures n is given. (1 ≤ n ≤ 1, 000, 000)
– In the second line, n start times s1, . . . , sn of the lectures are given in a line. (1 ≤
si ≤ 1, 999, 999)
– In the third line, n end times t1, . . . , tn of the lectures are given in the line. (si <
ti ≤ 2, 000, 000)