Flood
Rumi wants to develope a flood simulator. To do so, she made a grid map of Cheongpa-dong
and placed a number for each cell, so that 1 denotes a highland and 0 denotes a lowland. Now,
she wants to make an algorithm to count the number of flooded lowland cells if a flood occurs
from one lowland cell. Once flood is occured, it spreads through four neighboring cells until it
meets a highland cell or the end of the grid.
For instance, in the following 5 × 5 grid, if a flood occurs in the cell (2, 3) (red 0), 5 cells are
flooded (gray background)
 In the first line of the input, the number t of test cases is given.
 For each test case, the input is given as following.
– In the first line of a test case, the size of the grid n and m are given. (1 ≤ n, m ≤
2, 000)
– In the next following n lines, m numbers in each row of the grid is given, from top
to bottom.
– In the last line of a test case, the coordinates x and y of the lowland cell that a flood
occurs are given. (0 ≤ x < n, 0 ≤ y < m)