Maze Run
Minho woke up in a maze. The maze is represented as a grid, a cell containing 1 a corridor, and
a cell containing 0 indicates a wall. Teresa is now at the top-left corner and the exit is at the
bottom-right corner (the number of each corner is 1). Given a grid map of the maze, implement
an algorithm to compute the length of a shortest path from where Minho is to the exit.
For instance, in the following 5 × 5 grid, the shortest path has length 12 (pink line).
 In the first line of the input, the number t of test cases is given.
 For each test case, the input is given as follows.
– In the first line of a test case, the size (n × m) of the maze, n and m, is given.
(1 ≤ n, m ≤ 2, 000)
– In the next n lines, m numbers in each row are given, from top to bottom.